+++ src/Entities/Event.php
@@ -7,6 +7,7 @@ namespace Paddle\SDK\Entities;
 use Paddle\SDK\Entities\Event\EventTypeName;
 use Paddle\SDK\Notifications\Entities\Entity as NotificationEntity;
 use Paddle\SDK\Notifications\Entities\EntityFactory;
+use Paddle\SDK\Notifications\Entities\UndefinedEntity;
 use Paddle\SDK\Notifications\Events\UndefinedEvent;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -33,11 +34,16 @@ abstract class Event implements Entity
             $event = UndefinedEvent::class;
         }
 
+        // Create an undefined entity for undefined events.
+        $entity = $event === UndefinedEvent::class
+            ? UndefinedEntity::from($data['data'])
+            : EntityFactory::create($data['event_type'], $data['data']);
+
         return $event::fromEvent(
             $data['event_id'],
             EventTypeName::from($data['event_type']),
             DateTime::from($data['occurred_at']),
-            EntityFactory::create($data['event_type'], $data['data']),
+            $entity,
             $data['notification_id'] ?? null,
         );
     }